Second batch move from impala to base_sql
